Rename and refactor pipeline services
