+++ pipe/test_GammaFlash.sh
@@ -12,9 +12,9 @@ if [ "$#" -eq 0 ]; then
     exit 0
 fi
 # path_dl0='/home/worker/workspace/testgammaflash/Data/DL0/acquisizione_2022_06_24' 
-path_dl0='/home/usergamma/workspace/test/Data/DL0/prova.h5' 
-path_dl1='/home/usergamma/workspace/test/Data/DL1/prova.dl1.dl2.h5' 
-path_dl2='/home/usergamma/workspace/test/Data/DL2/prova.dl2.h5' 
+path_dl0='/home/usergamma/workspace/test/Data/DL0/' 
+path_dl1='/home/usergamma/workspace/test/Data/DL1/' 
+path_dl2='/home/usergamma/workspace/test/Data/DL2/' 
 path_json_result='/home/usergamma/workspace/test/Out'
 socket='tcp://127.0.0.1:5555'
 socket_cck='tcp://127.0.0.1:5559'
@@ -29,15 +29,15 @@ if [ $arg -eq 0 ]; then
 elif [ $arg -eq 1 ]; then
     echo ========= START DL0toDL2__service ========
     # rm -rf $path_dl2/*
-    python -c "from DL0toDL2__service.Supervisor_gflash import Supervisor_DL0toDL2; Supervisor_DL0toDL2('$config_path', 'DL0toDL2').start()"
+    python -c "from DL0toDL2__service.Supervisor_dl0todl2 import Supervisor_DL0toDL2; Supervisor_DL0toDL2('$config_path', 'DL0toDL2').start()"
 elif [ $arg -eq 2 ]; then
     echo ========= START DL0toDL1__service ========
     rm -rf /home/usergamma/workspace/test/Data/DL1/*
-    python -c "from DL0toDL1__service.Supervisor_gflash import Supervisor_DL0toDL1; Supervisor_DL0toDL1('${config_path}', 'DL0toDL1').start()"
+    python -c "from DL0toDL1__service.Supervisor_dl0todl1 import Supervisor_DL0toDL1; Supervisor_DL0toDL1('${config_path}', 'DL0toDL1').start()"
 elif [ $arg -eq 3 ]; then
     echo ========= START DL1toDL2__service ========
     rm -rf /home/usergamma/workspace/test/Data/DL2/*
-    python -c "from DL1toDL2__service.Supervisor_gflash import Supervisor_DL1toDL2; Supervisor_DL1toDL2('${config_path}', 'DL1toDL2').start()"
+    python -c "from DL1toDL2__service.Supervisor_dl1todl2 import Supervisor_DL1toDL2; Supervisor_DL1toDL2('${config_path}', 'DL1toDL2').start()"
 elif [ $arg -eq 4 ]; then
     echo ========= START DL2Checker ========
     # rm -rf $path_json_result/*
